fix loading first pic
